refactor: onHiddenChanged 내부 조건을 데이터를 성공적으로 불러온 적이 없을때로 수정
